refactor: move page methods to page.c
